+++ cirq-core/cirq/testing/__init__.py
@@ -109,8 +109,6 @@ from cirq.testing.routing_devices import (
 from cirq.testing.sample_circuits import nonoptimal_toffoli_circuit
 
 from cirq.testing.sample_gates import (
-    DecomposableGate,
-    NotDecomposableGate,
     GateThatAllocatesAQubit,
     GateThatAllocatesTwoQubits,
     GateThatDecomposesIntoNGates,
